Improved WindowExtensions. Decoupling logic code to make EscalateZBand reusable.
