+++ LenovoLegionToolkit.WPF/Windows/Osd/OsdWindowBase.cs
@@ -20,40 +20,13 @@ using LenovoLegionToolkit.Lib.Settings;
 using LenovoLegionToolkit.Lib.Utils;
 using LenovoLegionToolkit.WPF.Resources;
 using LenovoLegionToolkit.WPF.Settings;
+using LenovoLegionToolkit.WPF.Extensions;
 using WpfScreenHelper;
 
 namespace LenovoLegionToolkit.WPF.Windows.Osd;
 
 public abstract class OsdWindowBase : Window
 {
-    #region Win32
-
-    private const int GWL_EXSTYLE = -20;
-    private const int WS_EX_TRANSPARENT = 0x00000020;
-    private const int WS_EX_TOOLWINDOW = 0x00000080;
-    private const int WS_EX_NOACTIVATE = 0x08000000;
-
-    [DllImport("user32.dll")]
-    private static extern int GetWindowLong(IntPtr hWnd, int nIndex);
-
-    [DllImport("user32.dll")]
-    private static extern int SetWindowLong(IntPtr hWnd, int nIndex, int dwNewLong);
-
-    [DllImport("user32.dll", SetLastError = true)]
-    [return: MarshalAs(UnmanagedType.Bool)]
-    private static extern bool SetWindowPos(IntPtr hWnd, IntPtr hWndInsertAfter, int X, int Y, int cx, int cy, uint uFlags);
-
-    [DllImport("user32.dll", EntryPoint = "GetWindowBand")]
-    [return: MarshalAs(UnmanagedType.Bool)]
-    private static extern bool GetWindowBand(IntPtr hWnd, out uint pdwBand);
-
-    private static readonly IntPtr HWND_TOPMOST = new IntPtr(-1);
-    private const uint SWP_NOSIZE = 0x0001;
-    private const uint SWP_NOMOVE = 0x0002;
-    private const uint SWP_NOACTIVATE = 0x0010;
-
-    #endregion
-
     #region Threshold Constants
 
     protected int _uiUpdateThrottleMs = 0;
@@ -161,44 +134,7 @@ public abstract class OsdWindowBase : Window
 
     private void OnSourceInitialized(object? sender, EventArgs e)
     {
-        UpdateClickThrough();
-    }
-
-    private void UpdateClickThrough()
-    {
-        if (PresentationSource.FromVisual(this) is not HwndSource source) return;
-
-        var hwnd = source.Handle;
-        var extendedStyle = GetWindowLong(hwnd, GWL_EXSTYLE);
-
-        extendedStyle |= WS_EX_TOOLWINDOW | WS_EX_NOACTIVATE;
-
-        if (_OsdSettings.Store.IsLocked)
-            extendedStyle |= WS_EX_TRANSPARENT;
-        else
-            extendedStyle &= ~WS_EX_TRANSPARENT;
-
-        SetWindowLong(hwnd, GWL_EXSTYLE, extendedStyle);
-    }
-
-    private void EscalateZBand()
-    {
-        if (PresentationSource.FromVisual(this) is not HwndSource source) return;
-
-        var hwnd = source.Handle;
-        try
-        {
-            SetWindowPos(hwnd, HWND_TOPMOST, 0, 0, 0, 0, SWP_NOMOVE | SWP_NOSIZE | SWP_NOACTIVATE);
-
-            if (GetWindowBand(hwnd, out uint currentBand))
-            {
-                Log.Instance.Trace($"EscalateZBand (TOPMOST) executed. Current Band: {currentBand}");
-            }
-        }
-        catch (Exception ex)
-        {
-            Log.Instance.Trace($"Exception in EscalateZBand for HWND {hwnd:X}", ex);
-        }
+        this.SetClickThrough(_OsdSettings.Store.IsLocked);
     }
 
     private void OnMouseLeftButtonDown(object sender, MouseButtonEventArgs e)
@@ -251,13 +187,14 @@ public abstract class OsdWindowBase : Window
         }
     }
 
-    private void OnLoaded(object? sender, RoutedEventArgs e)
-        => Dispatcher.BeginInvoke(new Action(SetWindowPosition), DispatcherPriority.Loaded);
+    private void OnLoaded(object? sender, RoutedEventArgs e) => Dispatcher.BeginInvoke(new Action(SetWindowPosition), DispatcherPriority.Loaded);
 
     private void OnContentRendered(object? sender, EventArgs e)
     {
         if (!_positionSet)
+        {
             Dispatcher.BeginInvoke(new Action(SetWindowPosition), DispatcherPriority.Render);
+        }
     }
 
     protected virtual void SetWindowPosition()
@@ -328,7 +265,9 @@ public abstract class OsdWindowBase : Window
         Dispatcher.BeginInvoke(() =>
         {
             if (!IsPositionOnScreen(Left, Top))
+            {
                 SetDefaultWindowPosition();
+            }
         });
     }
 
@@ -387,7 +326,7 @@ public abstract class OsdWindowBase : Window
         _warningBrush = (Brush)converter.ConvertFromString(_OsdSettings.Store.WarningColor)!;
         _criticalBrush = (Brush)converter.ConvertFromString(_OsdSettings.Store.CriticalColor)!;
 
-        UpdateClickThrough();
+        this.SetClickThrough(_OsdSettings.Store.IsLocked);
 
         if (!SavedPositionX.HasValue || !SavedPositionY.HasValue)
         {
@@ -460,7 +399,7 @@ public abstract class OsdWindowBase : Window
         }
 
         CheckAndUpdateFpsMonitoring();
-        EscalateZBand();
+        this.EscalateZBand();
     }
 
     protected virtual void OnItemVisibilityChanged(FrameworkElement element, bool visible) { }
